% Auto-generated: do not edit by hand
\name{svg}

\alias{svg}

\title{Svg component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
svg(id=NULL, value=NULL, classMap=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks}

\item{value}{Character. The value displayed in the input}

\item{classMap}{Named list. A map of ids and the classes to be set on them for controlling
the inner svg element classes}
}
